# docs/swagger.yaml
basePath: /
definitions:
  handler.Album:
    properties:
      artist:
        example: John Coltrane
        type: string
      id:
        example: "1"
        type: string
      price:
        example: 56.99
        type: number
      title:
        example: Blue Train
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A sample album management API in Go.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Album API
  version: "1.0"
paths:
  /albums:
    get:
      description: Responds with the list of all albums as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Album'
            type: array
      summary: Get all albums
      tags:
      - albums
    post:
      consumes:
      - application/json
      description: Adds an album from JSON received in the request body.
      parameters:
      - description: Album JSON
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/handler.Album'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Album'
      summary: Post a new album
      tags:
      - albums
  /albums/{id}:
    get:
      description: Locates the album whose ID value matches the id parameter.
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Album'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an album by ID
      tags:
      - albums
  /ping:
    get:
      description: Responds with a simple "pong" message to test the server's responsiveness.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ping the server
      tags:
      - health
swagger: "2.0"